+++ packages/icons/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Breaking Change
 
--   Removed icons that were added by mistake: `alignJustifyAlt`, `cogAlt`.
+-   Removed icons that were added by mistake: `alignJustifyAlt`, `cogAlt`, `sparkles`, `trashFilled`. 
 
 ## 6.3.0 (2022-02-10)